Move WorkingOverview component to ControlOverview dir
