--- packages/airview-compliance-ui/src/features/working-overlay/index.js
@@ -1,1 +1,0 @@
-export { WorkingOverlay } from "./working-overlay";
